+++ gazebo/gui/GLWidget.cc
@@ -74,9 +74,9 @@ GLWidget::GLWidget(QWidget *_parent)
 
   this->dataPtr->windowId = -1;
 
-//   TODO causes problem inserting models and mouse picking in qt5
   this->setAttribute(Qt::WA_OpaquePaintEvent, true);
-//  this->setAttribute(Qt::WA_PaintOnScreen, true);
+  // This attribute causes problem with mouse picking in OSX/ogre1.9/qt5
+  // this->setAttribute(Qt::WA_PaintOnScreen, true);
 
   this->setFocusPolicy(Qt::StrongFocus);
   this->setMouseTracking(true);
@@ -237,7 +237,7 @@ void GLWidget::showEvent(QShowEvent *_event)
     // Get the window handle in a form that OGRE can use.
     std::string winHandle = this->OgreHandle();
 
-    // TODO windowhandle() is available in qt 5.0 only
+    // windowhandle() is available in qt 5.0 only
     double ratio = this->windowHandle()->devicePixelRatio();
 
     // Create the OGRE render window