+++ gazebo/physics/WorldPrivate.hh
@@ -27,6 +27,8 @@
 #include <sdf/sdf.hh>
 #include <string>
 
+#include <ignition/transport.hh>
+
 #include "gazebo/common/Event.hh"
 #include "gazebo/common/Time.hh"
 #include "gazebo/common/URI.hh"
@@ -344,6 +346,9 @@ namespace gazebo
 
       /// \brief All the introspection items regsitered for this.
       public: std::vector<common::URI> introspectionItems;
+
+      /// \brief Node for ignition transport communication.
+      public: ignition::transport::Node ignNode;
     };
   }
 }